Create firstconsolegame[Unfinished]
This is the first console "Game" i made when i had a discord tutor and she was amazed with what i did, anyways i dropped the project after i stopped learning because school had re started for 2-3 weeks and never did a lesson with her afterwars. 8/8 was my birthday and i started learning c# all over again on skillshare and i went back in this console game and made some changes somewhere at the end but i won't be continuing it so yeah.