Calculator
This ain't a wow project but i am still proud of myself of actually doing it! The only help i got was from my notebook. By the way on line 15, 16 it wouldn't let me print the message and the result at the same time, idk it would just print the "The final message is: " and stay like that