Calculator but completed
This is a completed version of the calculator with Subtract, multiply & divide. Got some help from the internet as i haven't learned about switches in my skillshare course but i think that's okay as long as i had the idea in the first place to try to get help. i also added notes.